export info getter and setter
