+++ crates/rspack_core/src/exports/mod.rs
@@ -4,6 +4,8 @@ mod export_info_setter;
 mod exports_info;
 mod exports_info_getter;
 mod exports_info_setter;
+mod referenced_export;
+mod target;
 mod utils;
 
 pub use export_info::*;
@@ -12,4 +14,6 @@ pub use export_info_setter::*;
 pub use exports_info::*;
 pub use exports_info_getter::*;
 pub use exports_info_setter::*;
+pub use referenced_export::*;
+pub use target::*;
 pub use utils::*;
